backend/app/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
firebase-admin==6.2.0
pydantic==2.0.3
python-dotenv==1.0.0
google-generativeai==0.7.2
httpx==0.27.2